move viewmodel abstractions to android-feature-viewmodel module
